Added few more elements
Lets start with link for main.css was not present on the index.html
page thats why my (dt) and (dd) were not working and I wasn’t able to
style the list. Next thing I did wrong is I wasn’t able to duplicate
the list because (writeberries2 [Blue, Grape, Red] ); was not present
this helps make the function understand that these are the three
variables were duplicating.